stdlib: get rid of uacpi_popcount
This made it so uACPI had a dependency on libgcc or the x86 instruction,
which is very inconvenient for the users of the library. Since it's used
in pretty much one single place, which needs it for up to 16 bits at max
anyway, let's switch to ad-hoc bit counting with a trivial for loop
instead of a general purpose popcount helper.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>