opcodes: add explicit properties parameter to UACPI_OP
Previously, the properties field was set by relying on the compiler to select
the next field in the struct after getting an explicit initializer for
decode_ops. However, some compilers - such as OpenWatcom - instead select
the next field in the union (indirect_decode_ops). Swapping the order of
these fields in the union is not sufficient, as the compiler refuses
to "leave" the union - instead, it results in error E1175 (Too many
initializers).