uacpi: work around OpenWatcom miscompilations
OpenWatcom miscompiles designated initializers that set fields within an
anonymous union member. To work around this, either the initializers for
these fields must be wrapped in braces (to indicate it's in a subobject),
or the anonymous union member must be named. This commit employs both
strategies; the former is preferred, but in the resource tables, it would
be an extremely invasive change due to the need to alter the parameter order
of nearly all OP() macro invocations.