interpreter: abort on most LoadOp errors
It's not a requirement to treat hard errors such as invalid input as
soft errors, which return False to the caller. Simply abort the method
in this case. The only error in handle_load we allow is a denied table
install, which we propagate as a soft error to the caller of Load.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>