tables: allow tables smaller than the ACPI SDT header
Apparently some firmware is capable of generating those, so let's treat
it as a warning. Note that this is only applicable for tables that come
directly from the firmware, we don't allow this for tables installed by
the user since it's impossible to deal with properly (for tables with a
virtual origin).

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>